+++ TWAdBlockAssetResourceLoaderDelegate.m
@@ -1,4 +1,5 @@
 #import "TWAdBlockAssetResourceLoaderDelegate.h"
+#import "SettingsKeys.h"
 
 extern NSUserDefaults *tweakDefaults;
 
@@ -11,9 +12,7 @@ extern NSUserDefaults *tweakDefaults;
   components.scheme = @"https";
   NSMutableURLRequest *request = loadingRequest.request.mutableCopy;
   request.URL = components.URL;
-  NSString *proxy = [tweakDefaults boolForKey:@"TWAdBlockCustomProxyEnabled"]
-                        ? [tweakDefaults stringForKey:@"TWAdBlockProxy"]
-                        : PROXY_ADDR;
+  NSString *proxy = twab_effectiveProxyAddress();
   NSURLSession *session = [[NSURLSession alloc] twab_proxySessionWithAddress:proxy];
   [[session dataTaskWithRequest:request
               completionHandler:^(NSData *data, NSURLResponse *response, NSError *error) {